# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quire Server API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "description": "Private reading-data sync and EPUB storage. Account and watched-folder administration are owner CLI operations."
  },
  "paths": {
...
    "/.well-known/quire": {
      "get": {
        "summary": "Discover this server",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "default": {
            "description": "Error; no batch changes committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "summary": "Create a 30-day device session",
        "security": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "default": {
            "description": "Error; no batch changes committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      },
      "get": {
        "summary": "List this account\u2019s active device sessions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    }
                  },
                  "required": [
                    "sessions"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; no batch changes committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{id}": {
      "delete": {
        "summary": "Revoke an owned device session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "default": {
            "description": "Error; no batch changes committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/sync": {
      "post": {
        "summary": "Atomically submit outbox operations and pull a page of changes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error; no batch changes committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          }
        },
        "description": "Request limit 2 MiB; at most 50 operations. All validated and committed together. Identical retries replay their original result. Stale revisions append a conflict candidate, including stale writes after deletion; current baseRevision explicitly resolves all candidates. More than 16 candidates, future revisions/cursors, or changed operation-ID content return 409 and roll back the entire batch. Each account has independent monotonic cursors. Pull at most 100 changes, repeat with returned cursor while hasMore. Never advance cursor to an operation result revision. Tombstones and retry history do not expire in v1."
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Discovery": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "apiVersion": {
            "const": "1"
          },
          "apiUrl": {
            "type": "string",
            "format": "uri"
          },
          "registration": {
            "const": "owner-only"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "apiVersion",
          "apiUrl",
          "registration",
          "capabilities"
        ]
      },
      "Login": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._-]{0,63}$"
          },
          "password": {
            "type": "string",
            "maxLength": 1024
          },
          "deviceName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "username",
          "password",
          "deviceName"
        ]
      },
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "expiresAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "deviceName",
          "createdAt",
          "expiresAt"
        ]
      },
      "LoginResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque bearer token. Returned once; store in OS credential storage."
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "required": [
          "token",
          "session"
        ]
      },
      "Book": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "author": {
            "type": "string",
            "maxLength": 2048
          },
          "series": {
            "type": "string",
            "maxLength": 2048
          },
          "volume": {
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "title"
        ]
      },
      "Position": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "cfi": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8192
          },
          "fraction": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "section": {
            "type": "string",
            "maxLength": 2048
          }
        },
        "required": [
          "cfi",
          "fraction"
        ]
      },
      "Annotation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "bookmark",
              "highlight"
            ]
          },
          "cfi": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8192
          },
          "text": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "section": {
            "type": "string",
            "maxLength": 2048
          }
        },
        "required": [
          "kind",
          "cfi"
        ]
      },
      "Operation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
          },
          "bookId": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "kind": {
            "enum": [
              "book",
              "position",
              "annotation"
            ]
          },
          "recordId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
            "description": "default for book/position; stable annotation ID otherwise."
          },
          "baseRevision": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "deleted": {
            "type": "boolean"
          },
          "value": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Book"
              },
              {
                "$ref": "#/components/schemas/Position"
              },
              {
                "$ref": "#/components/schemas/Annotation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "bookId",
          "kind",
          "recordId",
          "baseRevision",
          "deleted"
        ],
        "description": "Maximum 32 KiB encoded value. Value schema must match kind. Deleted operations omit value or use null. New records use baseRevision 0. Device timestamps are deliberately excluded. Operation IDs must not be reused for different content."
      },
      "Candidate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "operationId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
          },
          "deleted": {
            "type": "boolean"
          },
          "value": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Book"
              },
              {
                "$ref": "#/components/schemas/Position"
              },
              {
                "$ref": "#/components/schemas/Annotation"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "description": "Server Unix milliseconds."
          }
        },
        "required": [
          "operationId",
          "deleted",
          "value",
          "createdAt"
        ]
      },
...
      "Change": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "bookId": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "kind": {
            "enum": [
              "book",
              "position",
              "annotation"
            ]
          },
          "recordId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "candidates": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": {
              "$ref": "#/components/schemas/Candidate"
            }
          },
          "cursor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "bookId",
          "kind",
          "recordId",
          "revision",
          "candidates",
          "cursor"
        ]
      },
      "OperationResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "conflict": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "revision",
          "conflict"
        ]
      },
      "SyncRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "operations": {
            "type": [
              "array",
              "null"
            ],
            "maxItems": 50,
            "items": {
              "$ref": "#/components/schemas/Operation"
            }
          }
        },
        "required": []
      },
      "SyncResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OperationResult"
            }
          },
          "changes": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/Change"
            }
          },
          "cursor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "results",
          "changes",
          "cursor",
          "hasMore"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}